requirements.txt
Django==1.5.1
South==0.8.1
django-guardian==1.1.1
django-userena==1.2.1
easy-thumbnails==1.3
